Replace internal Stdin class with proper Stream handling
