+++ tests/StdioTest.php
@@ -20,7 +20,7 @@ class StdioTest extends TestCase
      */
     public function testCtorDefaultArgs()
     {
-        new Stdio($this->loop);
+        $stdio = new Stdio($this->loop);
 
         // Closing STDIN/STDOUT is not a good idea for reproducible tests
         // $stdio->close();
